Armorable now has a seperate damage mitigation / notification events
Also changed how the notifications worked
