+++ BoEG/Assets/Scripts/Framework/Core/Modules/Armorable/Armor.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 8e8108e3a6a5406e954d0fd6986c0bcb
+timeCreated: 1612828641
