+++ BoEG/Assets/Scripts/Framework/Core/Modules/Armorable/Armor/ArmorData.cs
