+++ BoEG/Assets/Scripts/Framework/Core/Modules/Armorable/Armor/ArmorX.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 4475f2f4dc7a47b09076094598783cfb
+timeCreated: 1612828632
