+++ BoEG/Assets/Scripts/Framework/Core/Modules/Armorable/Armor/IArmor.cs
@@ -1,0 +1,9 @@
+namespace MobaGame.Framework.Core.Modules
+{
+    public interface IArmor
+    {
+        float Block { get; }
+        float Resistance { get; }
+        bool Immunity { get; }
+    }
+}
