+++ BoEG/Assets/Scripts/Framework/Core/Modules/Armorable/Armor/IArmor.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 051eef75bdba4f778c4765c848b4148f
+timeCreated: 1612828632
