+++ BoEG/Assets/Scripts/Framework/Core/Modules/Armorable/Armor/IArmorView.cs
@@ -1,0 +1,10 @@
+namespace MobaGame.Framework.Core.Modules
+{
+    public interface IArmorView : IView
+    {
+        float Block { get; }
+        float Resistance { get; }
+        bool Immunity { get; }
+
+    }
+}
