+++ BoEG/Assets/Scripts/Framework/Core/Modules/Armorable/Armor/IArmorView.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: c8d4de71223f4b62a0ba5550b5e1fa81
+timeCreated: 1612828632
