+++ BoEG/Assets/Scripts/Framework/Core/Modules/Armorable/Armor/ModifiedArmor.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 04bc912965ec4b93b8803dd904496101
+timeCreated: 1612828632
