--- BoEG/Assets/Scripts/Framework/Core/Modules/Armorable/ArmorableEventArgs.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7b6e8ff947b5f5a41b2bfa91c800f3a7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
