+++ BoEG/Assets/Scripts/Framework/Core/Modules/Armorable/IArmorable.cs
@@ -3,26 +3,17 @@ using MobaGame.Framework.Types;
 
 namespace MobaGame.Framework.Core.Modules
 {
-    public interface IArmorableView : IView
-    {
-        IArmorView Physical { get; }
-        IArmorView Magical { get; }
-    }
     public interface IArmorable
     {
         IArmorableView View { get; }
         IArmor Physical { get; }
         IArmor Magical { get; }
-        Damage ResistDamage(Damage damage);
+        SourcedDamage ResistDamage(SourcedDamage damage);
 
-        float CalculateReduction(Damage damage);
+        float CalculateReduction(SourcedDamage damage);
 
-        event EventHandler<ArmorableEventArgs> Resisted;
-        event EventHandler<ArmorableEventArgs> Resisting;
+        event EventHandler<ChangedEventArgs<SourcedDamage>> Resisted;
+        event EventHandler<SourcedDamage> Resisting;
+        event EventHandler<ChangableEventArgs<SourcedDamage>> DamageMitigation;
     }
-    public static class IArmorableX
-    {
-        public static SourcedDamage ResistDamage(this IArmorable armorable, SourcedDamage damage) => damage.SetDamage(armorable.ResistDamage(damage.Value));
-    }
-
 }
