+++ BoEG/Assets/Scripts/Framework/Core/Modules/Armorable/IArmorableView.cs
@@ -1,0 +1,8 @@
+namespace MobaGame.Framework.Core.Modules
+{
+    public interface IArmorableView : IView
+    {
+        IArmorView Physical { get; }
+        IArmorView Magical { get; }
+    }
+}
