+++ BoEG/Assets/Scripts/Framework/Core/Modules/Armorable/IArmorableView.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 70a618fb7c324f6488ba8dbd89832c84
+timeCreated: 1612829315
