+++ BoEG/Assets/Scripts/Framework/Core/Modules/Armorable/Modifiers.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 159de5ba03a648cf9d6280b994ca7eb3
+timeCreated: 1612829344
