+++ BoEG/Assets/Scripts/Framework/Core/Modules/Armorable/Modifiers/IMagicalBlockModifier.cs
@@ -1,0 +1,7 @@
+namespace MobaGame.Framework.Core.Modules
+{
+    public interface IMagicalBlockModifier : IModifier
+    {
+        public FloatModifier MagicalBlock { get; }
+    }
+}
