+++ BoEG/Assets/Scripts/Framework/Core/Modules/Armorable/Modifiers/IMagicalBlockModifier.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 9bcefbf4280145f5ac1d2ddffe0575ba
+timeCreated: 1612829315
