+++ BoEG/Assets/Scripts/Framework/Core/Modules/Armorable/Modifiers/IMagicalResistanceModifier.cs
@@ -1,0 +1,7 @@
+namespace MobaGame.Framework.Core.Modules
+{
+    public interface IMagicalResistanceModifier : IModifier
+    {
+        public FloatModifier MagicalResistance { get; }
+    }
+}
