+++ BoEG/Assets/Scripts/Framework/Core/Modules/Armorable/Modifiers/IMagicalResistanceModifier.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 5982c81ee48044e7b5b378b9e0fee0dd
+timeCreated: 1612829315
