+++ BoEG/Assets/Scripts/Framework/Core/Modules/Armorable/Modifiers/IPhysicalBlockModifier.cs
@@ -1,0 +1,7 @@
+namespace MobaGame.Framework.Core.Modules
+{
+    public interface IPhysicalBlockModifier : IModifier
+    {
+        public FloatModifier PhysicalBlock { get; }
+    }
+}
