+++ BoEG/Assets/Scripts/Framework/Core/Modules/Armorable/Modifiers/IPhysicalResistanceModifier.cs
@@ -1,0 +1,7 @@
+namespace MobaGame.Framework.Core.Modules
+{
+    public interface IPhysicalResistanceModifier : IModifier
+    {
+        public FloatModifier PhysicalResistance { get; }
+    }
+}
