+++ BoEG/Assets/Scripts/Framework/Core/Modules/Armorable/Modifiers/IPhysicalResistanceModifier.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: fc3cf93ab14147529733d725318bc38e
+timeCreated: 1612829315
