+++ BoEG/Assets/Scripts/Framework/Core/Modules/Killable/DeathEventArgs.cs
@@ -4,9 +4,8 @@ namespace MobaGame.Framework.Core.Modules
 {
 	public class DeathEventArgs : EventArgs
     {
-        public DeathEventArgs(Actor actor) : this(actor, null) { }
 
-        public DeathEventArgs(Actor actor, Actor killer)
+        public DeathEventArgs(Actor actor, Actor killer = null)
         {
             Self = actor;
             Killer = killer;
