+++ BoEG/Assets/Scripts/Framework/Types/ChangableEventArgs.cs
