+++ BoEG/Assets/Scripts/Framework/Types/ChangableEventX.cs
