+++ BoEG/Assets/Scripts/Framework/Types/ChangedEventArgs.cs
