+++ BoEG/Assets/Scripts/Framework/Types/ChangedEventUtil.cs.meta
